requirements.txt
numpy==2.2.3
pillow==11.1.0
python-magic==0.4.27
python-magic-bin==0.4.14
